requirements.txt
# Core Dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Deep Learning / GenAI
tensorflow==2.13.0
keras==2.13.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.13.0


# Data Processing
pyyaml==6.0.1
tqdm==4.65.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing (optional)
pytest==7.4.0
